backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.117.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.9",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
